implement postMQl Repositories, services and their implementations

+++ src/org/mql/platform/dao/EducationRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * @author chermehdi
  */
-public interface EducationRepository extends JpaRepository<Education, Integer> {
+public interface EducationRepository extends JpaRepository<Education, Long> {
 
 }
