+++ src/org/mql/platform/dao/LaureateRepository.java
@@ -3,6 +3,6 @@ package org.mql.platform.dao;
 import org.mql.platform.models.Laureate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LaureateRepository extends JpaRepository<Laureate, Integer>{
+public interface LaureateRepository extends JpaRepository<Laureate, Long>{
 
 }
